Squashed commit of the following:
commit 61024a8eee60dace0e958c12ef31b35f10312ba8
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Fri Mar 15 17:27:05 2019 +0000

    rm EXTRA_VERBOSE

commit 80829dd0d87d69e2fdcb0ed5d0dc1b9847ebb82a
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Fri Mar 15 17:12:39 2019 +0000

    update test

commit c0f762d675396b67c13f98dcfaa067798a0151b1
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Fri Mar 15 17:03:35 2019 +0000

    unsigned long to uint32_t

commit 42970199e085e3c2511122c9b2e2a7dd2aa78f27
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Fri Mar 15 13:48:37 2019 +0000

    add extra verbose

commit 97e58116c64282b2aadd23c2b9edcc868ceeebfc
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 18:27:07 2019 +0000

    rm some lines

commit a2e7c3f8d5a27c7d9e123872a87ade24b4685de8
Merge: cfdbcf7 bf83f48
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 17:38:20 2019 +0000

    Merge branch 'master' into fix_appveyor

commit cfdbcf764636040140c6ce26c640ddeb13fd929a
Merge: 15e1803 5bf43b1
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 17:29:18 2019 +0000

    Merge branch 'master' into fix_appveyor

commit 15e1803e646ad066b2d3c71282438ea076d829f9
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 17:01:22 2019 +0000

    update path for gfortran

commit 01653272b6433c58bc098b0643568e03650ac643
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 13:52:16 2019 +0000

    fixing

commit e87cd45dedc9cfdb3a39612a9b948fa6b604762d
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 13:10:23 2019 +0000

    fixing

commit cdabaadecb083c41a35afc8717feb0f8760f5f42
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 13:07:33 2019 +0000

    fixing

commit 7c8906e7ee37bd8c14a9dcccd8c38ad6bb948b4b
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 13:01:59 2019 +0000

    fixing

commit 97dfc5b4a4fc79f9d5d465961c988b62a6a8264a
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 12:59:51 2019 +0000

    fixing

commit 7b243db83785ac921753a25c62ba15f245af19ed
Author: Brendan O'Donoghue <bodonoghue85@gmail.com>
Date:   Thu Mar 14 12:55:02 2019 +0000

    fixing

+++ src/rw.c
@@ -4,6 +4,7 @@
 #include "amatrix.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 /* writes/reads problem data to/from filename */
 /* This is a VERY naive implementation, doesn't care about portability etc */
@@ -118,39 +119,39 @@ static ScsData * read_scs_data(FILE *fin) {
 
 void SCS(write_data)(const ScsData *d, const ScsCone *k) {
   FILE* fout = fopen(d->stgs->write_data_filename, "wb");
-  unsigned long scs_int_sz = (unsigned long)sizeof(scs_int);
-  unsigned long scs_float_sz = (unsigned long)sizeof(scs_float);
+  uint32_t scs_int_sz = (uint32_t)sizeof(scs_int);
+  uint32_t scs_float_sz = (uint32_t)sizeof(scs_float);
   scs_printf("writing data to %s\n", d->stgs->write_data_filename);
-  fwrite(&(scs_int_sz), sizeof(unsigned long), 1, fout);
-  fwrite(&(scs_float_sz), sizeof(unsigned long), 1, fout);
+  fwrite(&(scs_int_sz), sizeof(uint32_t), 1, fout);
+  fwrite(&(scs_float_sz), sizeof(uint32_t), 1, fout);
   write_scs_cone(k, fout);
   write_scs_data(d, fout);
   fclose(fout);
 }
 
 scs_int SCS(read_data)(const char * filename, ScsData ** d, ScsCone ** k) {
-  unsigned long file_int_sz;
-  unsigned long file_float_sz;
+  uint32_t file_int_sz;
+  uint32_t file_float_sz;
   FILE* fin = fopen(filename, "rb");
   if (!fin) {
     scs_printf("Error reading file %s\n", filename);
     RETURN -1;
   }
   scs_printf("Reading data from %s\n", filename);
-  fread(&(file_int_sz), sizeof(unsigned long), 1, fin);
-  fread(&(file_float_sz), sizeof(unsigned long), 1, fin);
+  fread(&(file_int_sz), sizeof(uint32_t), 1, fin);
+  fread(&(file_float_sz), sizeof(uint32_t), 1, fin);
 
-  if (file_int_sz != (unsigned long)sizeof(scs_int)) {
+  if (file_int_sz != (uint32_t)sizeof(scs_int)) {
     scs_printf("Error, sizeof(file int) is %lu, but scs expects sizeof(int) "
         "%lu, scs should be recompiled with correct flags.\n",
-        file_int_sz, (unsigned long)sizeof(scs_int));
+        (unsigned long)file_int_sz, (unsigned long)sizeof(scs_int));
     fclose(fin);
     RETURN -1;
   }
-  if (file_float_sz != (unsigned long)sizeof(scs_float )) {
+  if (file_float_sz != (uint32_t)sizeof(scs_float )) {
     scs_printf("Error, sizeof(file float) is %lu, but scs expects "
         "sizeof(float) %lu, scs should be recompiled with the correct flags.\n",
-        file_float_sz, (unsigned long)sizeof(scs_float));
+        (unsigned long)file_float_sz, (unsigned long)sizeof(scs_float));
     fclose(fin);
     RETURN -1;
   }
